requirements.txt
fastapi==0.110.0
ipython==8.12.3
langchain==0.1.13
langchain_community==0.0.29
langchain_core==0.1.33
langchain_google_genai==0.0.11
langserve==0.0.51
matplotlib==3.8.3
Pillow==10.2.0
python-dotenv==1.0.1
Requests==2.31.0
uvicorn==0.29.0
